requirements.txt
# Core libraries
langchain
langchain-community
langchain-openai
openai
streamlit
pandas
numpy

# Data visualization
plotly
matplotlib

# API and data retrieval
tavily-python
yfinance
requests

# Utilities
python-dotenv
faiss-cpu
chromadb